Add DialWithDialContextFunc to add external dialer with context.
Before this change no context was passed to the external dialer. This
made it impossible to work out whether we should be re-using the TLS
session cache or not.

After this change we can store the tls Cache in use in the context and
re-use it in the external dialler as appropriate.